Added backwards compatibility with HDF5 subwell backends

+++ eu.openanalytics.phaedra.base.fs/src/eu/openanalytics/phaedra/base/fs/nio/NIOInterface.java
@@ -127,7 +127,7 @@ public class NIOInterface extends BaseFileServer {
 				}
 			});			
 		} else {
-			Files.delete(p);
+			if (Files.exists(p)) Files.delete(p);
 		}
 	}
